Update Pyst.py
Download the sounds folder and put it in the same folder of the project just like the repository, remember to install playsound==1.2.2 library in order to get the answer with a meme sound lmao. You can put all the sounds you want in .wav or .mp3 the code will randomly select a file no matter the number on the folder.